Configured router-dom to route to CpuDetail page
-Added Detail folder for Detail views
- Created Home page and moved logic from App.js
- CPU Glance in Home page links to separate CPU Detail page
