+++ sysmon-plus/src/Pages/Detail/DiskDetail.js
@@ -1,0 +1,13 @@
+// diskSection.js
+import React from 'react';
+
+const DiskDetail = () => {
+  return (
+    <div className="DiskDetail">
+      <h2>Disk</h2>
+      <p>**ENTER DISK NAME HERE**</p>
+    </div>
+  );
+};
+
+export default DiskDetail;
