+++ sysmon-plus/src/Pages/Detail/MemoryDetail.js
@@ -1,0 +1,13 @@
+// memSection.js
+import React from 'react';
+
+const MemoryDetail = () => {
+  return (
+    <div className="MemoryDetail">
+      <h2>Memory</h2>
+      <p>**ENTER MEMORY NAME HERE**</p>
+    </div>
+  );
+};
+
+export default MemoryDetail;
