--- sysmon-plus/src/Pages/DiskDetail.js
@@ -1,13 +1,0 @@
-// diskSection.js
-import React from 'react';
-
-const DiskDetail = () => {
-  return (
-    <div className="DiskDetail">
-      <h2>Disk</h2>
-      <p>**ENTER DISK NAME HERE**</p>
-    </div>
-  );
-};
-
-export default DiskDetail;
