--- sysmon-plus/src/Pages/MemoryDetail.js
@@ -1,13 +1,0 @@
-// memSection.js
-import React from 'react';
-
-const MemoryDetail = () => {
-  return (
-    <div className="MemoryDetail">
-      <h2>Memory</h2>
-      <p>**ENTER MEMORY NAME HERE**</p>
-    </div>
-  );
-};
-
-export default MemoryDetail;
